+++ lib/src/painting.dart
@@ -1,6 +1,7 @@
 const paintingSource = '''
 library painting;
 
+export 'package:flutter/src/painting/alignment.dart';
 export 'package:flutter/src/painting/basic_types.dart';
 export 'package:flutter/src/painting/edge_insets.dart';
 export 'package:flutter/src/painting/text_style.dart';