+++ lib/src/rendering.dart
@@ -1,3 +1,5 @@
 const renderingSource = '''
+export 'package:flutter/src/rendering/box.dart';
 export 'package:flutter/src/rendering/flex.dart';
+export 'package:flutter/src/rendering/object.dart';
 ''';