+++ lib/src/widgets/basic.dart
@@ -2,8 +2,10 @@ import 'package:dart_eval/dart_eval.dart';
 import 'package:dart_eval/dart_eval_bridge.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_eval/src/foundation/key.dart';
+import 'package:flutter_eval/src/painting/basic_types.dart';
 import 'package:flutter_eval/src/painting/edge_insets.dart';
 import 'package:flutter_eval/src/rendering/flex.dart';
+import 'package:flutter_eval/src/sky_engine/ui/text.dart';
 import 'package:flutter_eval/src/widgets/framework.dart';
 
 const widgetsBasicSource = '''
@@ -74,9 +76,11 @@ class $Column implements $Instance {
       constructors: {
         '': BridgeConstructorDef(BridgeFunctionDef(returns: BridgeTypeAnnotation($type), namedParams: [
           BridgeParameter('key', BridgeTypeAnnotation($Key.$type, nullable: true), true),
-          BridgeParameter('mainAxisSize', BridgeTypeAnnotation($MainAxisSize.$type), true),
           BridgeParameter('mainAxisAlignment', BridgeTypeAnnotation($MainAxisAlignment.$type), true),
+          BridgeParameter('mainAxisSize', BridgeTypeAnnotation($MainAxisSize.$type), true),
           BridgeParameter('crossAxisAlignment', BridgeTypeAnnotation($CrossAxisAlignment.$type), true),
+          BridgeParameter('verticalDirection', BridgeTypeAnnotation($VerticalDirection.$type), true),
+          BridgeParameter('textBaseline', BridgeTypeAnnotation($TextBaseline.$type), true),
           BridgeParameter(
               'children', BridgeTypeAnnotation(BridgeTypeRef.type(RuntimeTypes.listType, [$Widget.$type])), true),
         ]))
@@ -92,10 +96,12 @@ class $Column implements $Instance {
   static $Value? $new(Runtime runtime, $Value? target, List<$Value?> args) {
     return $Column.wrap(Column(
       key: args[0]?.$value,
-      mainAxisSize: args[1]?.$value ?? MainAxisSize.max,
-      mainAxisAlignment: args[2]?.$value ?? MainAxisAlignment.start,
+      mainAxisAlignment: args[1]?.$value ?? MainAxisAlignment.start,
+      mainAxisSize: args[2]?.$value ?? MainAxisSize.max,
       crossAxisAlignment: args[3]?.$value ?? CrossAxisAlignment.center,
-      children: (args[4]?.$reified as List).cast(),
+      verticalDirection: args[4]?.$value ?? VerticalDirection.down,
+      textBaseline: args[5]?.$value,
+      children: (args[6]?.$reified as List).cast(),
     ));
   }
 
@@ -126,9 +132,11 @@ class $Row implements $Instance {
       constructors: {
         '': BridgeConstructorDef(BridgeFunctionDef(returns: BridgeTypeAnnotation($type), namedParams: [
           BridgeParameter('key', BridgeTypeAnnotation($Key.$type, nullable: true), true),
-          BridgeParameter('mainAxisSize', BridgeTypeAnnotation($MainAxisSize.$type), true),
           BridgeParameter('mainAxisAlignment', BridgeTypeAnnotation($MainAxisAlignment.$type), true),
+          BridgeParameter('mainAxisSize', BridgeTypeAnnotation($MainAxisSize.$type), true),
           BridgeParameter('crossAxisAlignment', BridgeTypeAnnotation($CrossAxisAlignment.$type), true),
+          BridgeParameter('textDirection', BridgeTypeAnnotation($TextDirection.$type), true),
+          BridgeParameter('verticalDirection', BridgeTypeAnnotation($VerticalDirection.$type), true),
           BridgeParameter(
               'children', BridgeTypeAnnotation(BridgeTypeRef.type(RuntimeTypes.listType, [$Widget.$type])), true),
         ]))
@@ -144,10 +152,12 @@ class $Row implements $Instance {
   static $Value? $new(Runtime runtime, $Value? target, List<$Value?> args) {
     return $Row.wrap(Row(
       key: args[0]?.$value,
-      mainAxisSize: args[1]?.$value ?? MainAxisSize.max,
-      mainAxisAlignment: args[2]?.$value ?? MainAxisAlignment.start,
+      mainAxisAlignment: args[1]?.$value ?? MainAxisAlignment.start,
+      mainAxisSize: args[2]?.$value ?? MainAxisSize.max,
       crossAxisAlignment: args[3]?.$value ?? CrossAxisAlignment.center,
-      children: (args[4]?.$reified as List).cast(),
+      textDirection: args[4]?.$value,
+      verticalDirection: args[5]?.$value ?? VerticalDirection.down,
+      children: (args[6]?.$reified as List).cast(),
     ));
   }
 
@@ -178,6 +188,8 @@ class $Center implements $Instance {
       constructors: {
         '': BridgeConstructorDef(BridgeFunctionDef(returns: BridgeTypeAnnotation($type), namedParams: [
           BridgeParameter('key', BridgeTypeAnnotation($Key.$type, nullable: true), true),
+          BridgeParameter('widthFactor', BridgeTypeAnnotation(BridgeTypeRef.type(RuntimeTypes.doubleType)), true),
+          BridgeParameter('heightFactor', BridgeTypeAnnotation(BridgeTypeRef.type(RuntimeTypes.doubleType)), true),
           BridgeParameter('child', BridgeTypeAnnotation($Widget.$type), true),
         ]))
       },
@@ -192,7 +204,9 @@ class $Center implements $Instance {
   static $Value? $new(Runtime runtime, $Value? target, List<$Value?> args) {
     return $Center.wrap(Center(
       key: args[0]?.$value,
-      child: args[1]?.$value,
+      widthFactor: args[1]?.$value,
+      heightFactor: args[2]?.$value,
+      child: args[3]?.$value,
     ));
   }
 
@@ -201,6 +215,51 @@ class $Center implements $Instance {
 
   @override
   Center get $reified => $value;
+
+  @override
+  $Value? $getProperty(Runtime runtime, String identifier) {
+    throw UnimplementedError();
+  }
+
+  @override
+  int get $runtimeType => throw UnimplementedError();
+
+  @override
+  void $setProperty(Runtime runtime, String identifier, $Value value) {
+    throw UnimplementedError();
+  }
+}
+
+class $Builder implements $Instance {
+  static const $type = BridgeTypeRef.spec(BridgeTypeSpec('package:flutter/src/widgets/basic.dart', 'Builder'));
+
+  static const $declaration = BridgeClassDef(BridgeClassType($type, isAbstract: false, $extends: $Widget.$type),
+      constructors: {
+        '': BridgeConstructorDef(BridgeFunctionDef(returns: BridgeTypeAnnotation($type), namedParams: [
+          BridgeParameter('key', BridgeTypeAnnotation($Key.$type, nullable: true), true),
+          BridgeParameter('builder', BridgeTypeAnnotation(BridgeTypeRef.type(RuntimeTypes.functionType)), true),
+        ]))
+      },
+      methods: {},
+      getters: {},
+      setters: {},
+      fields: {},
+      wrap: true);
+
+  const $Builder.wrap(this.$value);
+
+  static $Value? $new(Runtime runtime, $Value? target, List<$Value?> args) {
+    return $Builder.wrap(Builder(
+      key: args[0]?.$value,
+      builder: args[1]?.$value,
+    ));
+  }
+
+  @override
+  final Builder $value;
+
+  @override
+  Builder get $reified => $value;
 
   @override
   $Value? $getProperty(Runtime runtime, String identifier) {