+++ lib/src/widgets/container.dart
@@ -3,7 +3,9 @@ import 'package:dart_eval/dart_eval_bridge.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_eval/src/foundation/key.dart';
+import 'package:flutter_eval/src/painting/alignment.dart';
 import 'package:flutter_eval/src/painting/edge_insets.dart';
+import 'package:flutter_eval/src/rendering/box.dart';
 import 'package:flutter_eval/src/sky_engine/ui/painting.dart';
 
 import 'framework.dart';
@@ -18,13 +20,17 @@ class $Container implements Container, $Instance {
       constructors: {
         '': BridgeConstructorDef(BridgeFunctionDef(returns: BridgeTypeAnnotation($type), namedParams: [
           BridgeParameter('key', BridgeTypeAnnotation($Key.$type, nullable: true), true),
-          BridgeParameter('color', BridgeTypeAnnotation($Color.$type, nullable: true), true),
-          BridgeParameter('child', BridgeTypeAnnotation($Widget.$type, nullable: true), true),
+          BridgeParameter('alignment', BridgeTypeAnnotation($AlignmentGeometry.$type, nullable: true), true),
           BridgeParameter('padding', BridgeTypeAnnotation($EdgeInsetsGeometry.$type, nullable: true), true),
+          BridgeParameter('color', BridgeTypeAnnotation($Color.$type, nullable: true), true),
           BridgeParameter(
               'width', BridgeTypeAnnotation(BridgeTypeRef.type(RuntimeTypes.doubleType), nullable: true), true),
           BridgeParameter(
               'height', BridgeTypeAnnotation(BridgeTypeRef.type(RuntimeTypes.doubleType), nullable: true), true),
+          BridgeParameter('constraints', BridgeTypeAnnotation($BoxConstraints.$type, nullable: true), true),
+          BridgeParameter('margin', BridgeTypeAnnotation($EdgeInsetsGeometry.$type, nullable: true), true),
+          BridgeParameter('transformAlignment', BridgeTypeAnnotation($AlignmentGeometry.$type, nullable: true), true),
+          BridgeParameter('child', BridgeTypeAnnotation($Widget.$type, nullable: true), true),
         ]))
       },
       methods: {},
@@ -36,11 +42,15 @@ class $Container implements Container, $Instance {
   static $Value? $new(Runtime runtime, $Value? target, List<$Value?> args) {
     return $Container.wrap(Container(
       key: args[0]?.$value,
-      color: args[1]?.$value,
-      child: args[2]?.$value,
-      padding: args[3]?.$value,
+      alignment: args[1]?.$value,
+      padding: args[2]?.$value,
+      color: args[3]?.$value,
       width: args[4]?.$value,
       height: args[5]?.$value,
+      constraints: args[6]?.$value,
+      margin: args[7]?.$value,
+      transformAlignment: args[8]?.$value,
+      child: args[9]?.$value,
     ));
   }